fix issue and add tests
